[DH] Spot Comments / UPDATE Spot Comment CRUD

+++ frontend/src/components/spots/SpotCard.vue
@@ -48,7 +48,7 @@
 
       <v-divider style="background-color:#FF5E5E;"></v-divider>
 
-      <div class="modalbody" style="height: 28vw;">
+      <div class="modalbody" style="height: 30vw;">
         <v-container>
           <v-row>
             <v-col cols="6">
@@ -148,20 +148,34 @@
     },
     methods: {
       likespot() {
-        axios.post(process.env.VUE_APP_SERVER_URL + SERVER.URL.SPOT.SPOTS + this.spot.id + SERVER.URL.SPOT.LIKE, null, this.headers)
-        .then(() => {
-          this.spot.is_liked = true;
-          this.spot.total_likes += 1;
-        })
-        .catch(err => console.log(err.response))
+        if (!this.LoginFlag) {
+          let response = confirm('로그인이 필요한 기능입니다!\n\n로그인 페이지로 이동하시겠습니까?')
+          if (response) {
+            this.$router.push({name: 'UsersLogin'})
+          }
+        } else {
+          axios.post(process.env.VUE_APP_SERVER_URL + SERVER.URL.SPOT.SPOTS + this.spot.id + SERVER.URL.SPOT.LIKE, null, this.headers)
+          .then(() => {
+            this.spot.is_liked = true;
+            this.spot.total_likes += 1;
+          })
+          .catch(err => console.log(err.response))
+        }
       },
       unlikespot() {
-        axios.delete(process.env.VUE_APP_SERVER_URL + SERVER.URL.SPOT.SPOTS + this.spot.id + SERVER.URL.SPOT.LIKE, this.headers)
-        .then(() => {
-          this.spot.is_liked = false;
-          this.spot.total_likes -= 1;
-        })
-        .catch(err => console.log(err.response))
+        if (!this.LoginFlag) {
+          let response = confirm('로그인이 필요한 기능입니다!\n\n로그인 페이지로 이동하시겠습니까?')
+          if (response) {
+            this.$router.push({name: 'UsersLogin'})
+          }
+        } else {
+          axios.delete(process.env.VUE_APP_SERVER_URL + SERVER.URL.SPOT.SPOTS + this.spot.id + SERVER.URL.SPOT.LIKE, this.headers)
+          .then(() => {
+            this.spot.is_liked = false;
+            this.spot.total_likes -= 1;
+          })
+          .catch(err => console.log(err.response))
+        }
       },
     },
   }
